# loadbalancer: Optimize ServiceName and L3n4Addr key construction