--- a/pkg/ciliumenvoyconfig/testdata/marshalling.txtar
+++ b/pkg/ciliumenvoyconfig/testdata/marshalling.txtar
@@ -366,11 +366,7 @@ spec:
   },
   "Redirects": [
     {
-      "k": {
-        "Namespace": "test",
-        "Name": "echo",
-        "Cluster": ""
-      },
+      "k": "test/echo",
       "v": {
         "ProxyPort": 1000,
         "Ports": [
@@ -381,11 +377,7 @@ spec:
   ],
   "ReconciledRedirects": [
     {
-      "k": {
-        "Namespace": "test",
-        "Name": "echo",
-        "Cluster": ""
-      },
+      "k": "test/echo",
       "v": {
         "ProxyPort": 1000,
         "Ports": [
@@ -395,10 +387,6 @@ spec:
     }
   ],
   "ReferencedServices": [
-    {
-      "Namespace": "test",
-      "Name": "echo",
-      "Cluster": ""
-    }
+    "test/echo"
   ]
 }