--- a/pkg/loadbalancer/loadbalancer.go
+++ b/pkg/loadbalancer/loadbalancer.go
@@ -5,6 +5,7 @@ package loadbalancer
 
 import (
 	"encoding/binary"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"net"
@@ -12,13 +13,16 @@ import (
 	"slices"
 	"strconv"
 	"strings"
+	"unsafe"
 
+	"github.com/cespare/xxhash/v2"
 	"github.com/cilium/statedb/index"
 	"github.com/cilium/statedb/part"
 	"go.yaml.in/yaml/v3"
 
 	"github.com/cilium/cilium/api/v1/models"
 	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
+	"github.com/cilium/cilium/pkg/container/cache"
 	"github.com/cilium/cilium/pkg/hive"
 	"github.com/cilium/cilium/pkg/u8proto"
 )
@@ -480,53 +484,165 @@ type ServiceID uint16
 // including both the namespace and name of the service (and optionally the cluster).
 // +k8s:deepcopy-gen=true
 type ServiceName struct {
-	Namespace string
-	Name      string
-	Cluster   string
+	// str is (<cluster>/)<namespace>/<name>
+	str string
+
+	// namePos is where the name starts
+	// (<cluster>/)<namespace>/<name>
+	//                         ^
+	namePos uint16
+
+	// clusterEndPos is where the cluster (including '/' ends. If zero then there is
+	// no cluster.
+	// (<cluster>/)<namespace>/<name>
+	//             ^
+	clusterEndPos uint16
 }
 
-func (m ServiceName) MarshalYAML() (any, error) {
-	return m.String(), nil
-}
-
-func (m *ServiceName) UnmarshalYAML(value *yaml.Node) error {
-	parts := strings.Split(value.Value, "/")
-	switch len(parts) {
-	case 0: /* empty name */
-	case 1:
-		m.Name = parts[0]
-	case 2:
-		m.Namespace = parts[0]
-		m.Name = parts[1]
-	case 3:
-		m.Cluster = parts[0]
-		m.Namespace = parts[1]
-		m.Name = parts[2]
-	default:
-		return fmt.Errorf("expected 0, 1 or 2 slashes, got %d", len(parts))
+func (s ServiceName) Cluster() string {
+	if s.clusterEndPos > 0 {
+		return s.str[:s.clusterEndPos-1]
 	}
+	return ""
+}
+
+func (s ServiceName) Name() string {
+	return s.str[s.namePos:]
+}
+
+func (s ServiceName) Namespace() string {
+	return s.str[s.clusterEndPos : s.namePos-1]
+}
+
+func (n ServiceName) Key() index.Key {
+	// index.Key is never mutated so it's safe to return the underlying
+	// string as []byte without copying.
+	return unsafe.Slice(unsafe.StringData(n.str), len(n.str))
+}
+
+func NewServiceName(namespace, name string) ServiceName {
+	return NewServiceNameInCluster("", namespace, name)
+}
+
+// serviceNameCache for deduplicating the [ServiceName.str] to reduce overall
+// memory usage.
+var serviceNameCache = cache.New(
+	func(n ServiceName) uint64 {
+		return serviceNameHash(n.Cluster(), n.Namespace(), n.Name())
+	},
+	nil,
+	func(a, b ServiceName) bool {
+		return b.str != "" /* only match if non-zero value */ &&
+			a.str == b.str
+	},
+)
+
+func serviceNameHash(cluster, namespace, name string) uint64 {
+	var d xxhash.Digest
+	d.WriteString(cluster)
+	d.WriteString(namespace)
+	d.WriteString(name)
+	return d.Sum64()
+}
+
+func NewServiceNameInCluster(cluster, namespace, name string) ServiceName {
+	return cache.GetOrPutWith(
+		serviceNameCache,
+		serviceNameHash(cluster, namespace, name),
+		func(sn ServiceName) bool {
+			return len(sn.str) > 0 &&
+				sn.Cluster() == cluster && sn.Namespace() == namespace && sn.Name() == name
+		},
+		func() ServiceName {
+			// ServiceName not found from cache, create it.
+			var b strings.Builder
+			pos := 0
+			if cluster != "" {
+				n, _ := b.WriteString(cluster)
+				b.WriteRune('/')
+				pos += n + 1
+			}
+			cendPos := pos
+			n, _ := b.WriteString(namespace)
+			b.WriteRune('/')
+			pos += n + 1
+			b.WriteString(name)
+			return ServiceName{
+				str:           b.String(),
+				clusterEndPos: uint16(cendPos),
+				namePos:       uint16(pos),
+			}
+		},
+	)
+}
+
+func (n ServiceName) MarshalJSON() ([]byte, error) {
+	return json.Marshal(n.str)
+}
+
+func (n *ServiceName) UnmarshalJSON(bs []byte) error {
+	return n.unmarshalString(strings.Trim(string(bs), `"`))
+}
+
+func (n *ServiceName) unmarshalString(s string) error {
+	s = strings.TrimSpace(s[:min(len(s), 65535)])
+	orig := s
+	n.str = s
+	pos := 0
+	popSlash := func() int {
+		if len(s) > 0 {
+			idx := strings.Index(s, "/")
+			if idx >= 0 {
+				s = s[idx+1:]
+				pos += idx + 1
+				return pos
+			}
+		}
+		return -1
+	}
+	i1, i2 := popSlash(), popSlash()
+	switch {
+	case i1 < 0:
+		n.str = ""
+		return fmt.Errorf("invalid service name: no namespace in %q", orig)
+	case i2 < 0:
+		n.namePos = uint16(i1)
+	default:
+		n.clusterEndPos = uint16(i1)
+		n.namePos = uint16(i2)
+	}
+	// Deduplicate
+	*n = serviceNameCache.Get(*n)
 	return nil
 }
 
+func (n ServiceName) MarshalYAML() (any, error) {
+	return n.String(), nil
+}
+
+func (n *ServiceName) UnmarshalYAML(value *yaml.Node) error {
+	return n.unmarshalString(value.Value)
+}
+
 func (n *ServiceName) Equal(other ServiceName) bool {
-	return n.Namespace == other.Namespace &&
-		n.Name == other.Name &&
-		n.Cluster == other.Cluster
+	return n.clusterEndPos == other.clusterEndPos &&
+		n.namePos == other.namePos &&
+		n.str == other.str
 }
 
 func (n ServiceName) Compare(other ServiceName) int {
 	switch {
-	case n.Namespace < other.Namespace:
+	case n.Namespace() < other.Namespace():
 		return -1
-	case n.Namespace > other.Namespace:
+	case n.Namespace() > other.Namespace():
 		return 1
-	case n.Name < other.Name:
+	case n.Name() < other.Name():
 		return -1
-	case n.Name > other.Name:
+	case n.Name() > other.Name():
 		return 1
-	case n.Cluster < other.Cluster:
+	case n.Cluster() < other.Cluster():
 		return -1
-	case n.Cluster > other.Cluster:
+	case n.Cluster() > other.Cluster():
 		return 1
 	default:
 		return 0
@@ -534,19 +650,12 @@ func (n ServiceName) Compare(other ServiceName) int {
 }
 
 func (n ServiceName) String() string {
-	if n.Cluster != "" {
-		return n.Cluster + "/" + n.Namespace + "/" + n.Name
-	}
-
-	return n.Namespace + "/" + n.Name
+	return n.str
 }
 
 func (n ServiceName) AppendSuffix(suffix string) ServiceName {
-	return ServiceName{
-		Namespace: n.Namespace,
-		Name:      n.Name + suffix,
-		Cluster:   n.Cluster,
-	}
+	n.str += suffix
+	return n
 }
 
 // BackendID is the backend's ID.
@@ -1047,6 +1156,6 @@ func NewL3n4AddrFromBackendModel(base *models.BackendAddress) (*L3n4Addr, error)
 func init() {
 	// Register the types for use with part.Map and part.Set.
 	part.RegisterKeyType(
-		func(name ServiceName) []byte { return []byte(name.String()) })
+		func(name ServiceName) []byte { return []byte(name.Key()) })
 	part.RegisterKeyType(L3n4Addr.Bytes)
 }