--- a/pkg/loadbalancer/redirectpolicy/cell.go
+++ b/pkg/loadbalancer/redirectpolicy/cell.go
@@ -8,7 +8,6 @@ import (
 	"github.com/cilium/statedb"
 
 	"github.com/cilium/cilium/api/v1/server/restapi/service"
-	"github.com/cilium/cilium/pkg/k8s"
 	"github.com/cilium/cilium/pkg/loadbalancer"
 	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/metrics/metric"
@@ -94,17 +93,17 @@ func newControllerMetrics() controllerMetrics {
 }
 
 type LRPMetrics interface {
-	AddLRPConfig(k8s.ServiceID)
-	DelLRPConfig(k8s.ServiceID)
+	AddLRPConfig(loadbalancer.ServiceName)
+	DelLRPConfig(loadbalancer.ServiceName)
 }
 
 type lrpMetricsNoop struct {
 }
 
-func (p *lrpMetricsNoop) AddLRPConfig(k8s.ServiceID) {
+func (p *lrpMetricsNoop) AddLRPConfig(loadbalancer.ServiceName) {
 }
 
-func (p *lrpMetricsNoop) DelLRPConfig(k8s.ServiceID) {
+func (p *lrpMetricsNoop) DelLRPConfig(loadbalancer.ServiceName) {
 }
 
 func NewLRPMetricsNoop() LRPMetrics {