* Use weak.Pointer for container/cache to avoid ending up with many cache arrays when there are lots of cores. Since we're doing a lot of string dedup'ing already this should increase cache hit rates and reduce memory usage overall in the agent. Since we'll only hold onto one array, bump the cache size to 1024 entries to improve hit rate (8*1024=~8k per cache, and we have Strings, StringMap, ServiceCache and L3n4AddrBytes caches, so static allocation of ~32k for them all).
* Use a single string for ServiceName allowing direct casting into the StateDB key
* Use a cache for L3n4Addr Bytes() to greatly reduce constructing the StateDB from scratch
* Remove some old unused methods from `loadbalancer.go`
* Remove `k8s.ServiceID` and use `loadbalancer.ServiceName` instead in `redirectpolicy` and `k8s.Endpoints`

Before:
```
pkg: github.com/cilium/cilium/pkg/loadbalancer/writer
cpu: 13th Gen Intel(R) Core(TM) i9-13950HX
Benchmark_UpsertServiceAndFrontends_100-32          2673            424350 ns/op            235655 objects/sec    456719 B/op       6286 allocs/op
Avg: Allocated 685859kB in total, 2441351 objects / 166553kB still reachable (per service:  48 objs, 14046B alloc,  3411B in-use)
Avg: Reconciled 50000 objects in 846.8988ms  (16.937µs  per service /  59042 services per second)
```

After:
```
Benchmark_UpsertServiceAndFrontends_100-32          2842            408752 ns/op            244647 objects/sec    428990 B/op       5202 allocs/op
Avg: Allocated 646749kB in total, 2381278 objects / 162073kB still reachable (per service:  47 objs, 13245B alloc,  3319B in-use)
Avg: Reconciled 50000 objects in 831.240522ms (16.624µs  per service /  60154 services per second)

```

`Benchmark_UpsertServiceAndFrontends` allocates ~15% less, ~4% faster.
~6% less allocations and ~4% less resident memory use in `pkg/loadbalancer/benchmark`.